Fixed database connections and genreralized code to handle multiple coins.
